+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 build
 src/data.json
+src/data-rare.json
